[ci-skip] Fix various warnings. (#878)
Small fix to Tasks.h ( incorrect override )
Sprinkling `final` on a few serialization functions.

+++ Projects/CoX/Common/GameData/Entity.h
@@ -216,7 +216,7 @@ public:
 
         std::array<PosUpdate, 64> m_pos_updates;
         std::array<BinTreeEntry, 7> m_interp_bintree;
-        size_t              m_update_idx                = 0;
+        int                 m_update_idx                = 0;
         bool                m_pchar_things              = false;
         bool                m_update_anims              = false;
         bool                m_hasname                   = false;
