+++ Projects/CoX/Common/GameData/GameDataStore.cpp
@@ -344,10 +344,10 @@ uint32_t GameDataStore::expDebtForLevel(uint32_t lev) const
 uint32_t GameDataStore::expMaxLevel() const
 {
     // return -1 because level is stored in indexed array (starting 0)
-    return m_experience_and_debt_per_level.m_ExperienceRequired.size()-1;
+    return uint32_t(m_experience_and_debt_per_level.m_ExperienceRequired.size()-1);
 }
 
-int GameDataStore::countForLevel(uint32_t lvl, const std::vector<uint32_t> &schedule) const
+uint32_t GameDataStore::countForLevel(uint32_t lvl, const std::vector<uint32_t> &schedule) const
 {
     uint32_t i = 0;
